src/requirements.txt
fastapi
arq
starlette
openai
python-dotenv
uvicorn
jinja2
markdown
python-docx
docx
python-multipart
json_repair